src/requirements.txt
fastapi
pyjwt
bcrypt
sqlalchemy
psycopg2
pydantic[email]
uvicorn[standard]
